Rewrite the ``@dr.freeze`` frontend on top of the new ``Layout`` class
The frontend used to rebuild a flattened description of the input on
every call, hash it, probe a map of recordings, and after replaying walk
the input a second time to write every output back, including the ones
that had not changed.

This commit rewrites ``freeze.cpp`` on top of the ``Layout`` class from
the previous commit. A call now proceeds as follows:

- Fast path: walk through the last used recording to verify the input
  for compatibility and simultaneously collect the deduplicated JIT
  variable indices. This is done with minimal use of the CPython API
  and without C++-side dynamic allocation.

- Only if that fails, build a layout and search the cache, which is now a
  vector compared with ``layout_equal()``. Record if nothing matches.

- After replay, construct the result from the output layout and write
  back only the modified input leaves.

Recording is stricter in one regard: a function may assign new arrays to
its input, but any other change to it now raises an error naming the
entry that changed. A dictionary key or attribute added by the function
used to be reconstructed from the recording on every replay, and a
changed tensor shape (the rank or a trailing dimension, since the
leading one is not part of the recording) was silently dropped. The
commit also removes the ``jit_freeze_discard()`` that added some
complexity.

Diagnostics now refer to function inputs by parameter name. Warnings about
repeated tracing quote what changed compared to the previous recording.

+++ src/extra/texture.cpp
@@ -223,7 +223,7 @@ struct JitOps {
         ad_loop_write write_cb = [](void *q, const dr::vector<uint64_t> &indices,
                                     bool) {
             Payload *pl = (Payload *) q;
-            pl->i = Int::borrow(indices[0]);
+            pl->i = Int::borrow((uint32_t) indices[0]);
             for (uint32_t j = 0; j < pl->n_state; ++j)
                 pl->state[j] = Float::borrow(indices[j + 1]);
         };
